% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allmap}
\alias{allmap}
\title{map from biocViews to EDAM}
\format{
An object of class \code{data.frame} with 917 rows and 7 columns.
}
\usage{
allmap
}
\description{
map from biocViews to EDAM
}
\note{
uses pypi text2term
}
\keyword{datasets}
