% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_bioc.R
\name{bioc_line}
\alias{bioc_line}
\title{use BiocPkgTools to acquire metadata on a specified Bioconductor package}
\usage{
bioc_line(pkgname = "tximeta")
}
\arguments{
\item{pkgname}{character(1)}
}
\value{
a 1-line tibble with all the fields defined in BiocPkgTools::biocPkgList output
}
\description{
use BiocPkgTools to acquire metadata on a specified Bioconductor package
}
