% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biotools_bioc}
\alias{biotools_bioc}
\title{result of scanning biotools metadata on 3 Nov 2024 to find bioconductor packages and EDAM topics}
\format{
An object of class \code{data.frame} with 3364 rows and 5 columns.
}
\usage{
biotools_bioc
}
\description{
result of scanning biotools metadata on 3 Nov 2024 to find bioconductor packages and EDAM topics
}
\note{
Used R code to scrape `https://github.com/research-software-ecosystem/content` to obtain this table.
See code in 'inst/scrapes' folder in biocEDAM package.
}
\keyword{datasets}
