% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvbrowse.R
\name{bvbrowse}
\alias{bvbrowse}
\title{interactive exploration of biocViews in relation to packages}
\usage{
bvbrowse()
}
\value{
called for its side effect (launches a Shiny app);
returns \code{NULL} invisibly when the app is stopped
}
\description{
interactive exploration of biocViews in relation to packages
}
\note{
Was made consistent with inst/app code on 3 Nov 2024.
}
\examples{
if (interactive()) bvbrowse()
}
