% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvbrowse.R
\name{bvserver}
\alias{bvserver}
\title{interactive exploration of biocViews in relation to packages}
\usage{
bvserver(input, output)
}
\description{
interactive exploration of biocViews in relation to packages
}
