% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edamize.R
\name{cleantxt}
\alias{cleantxt}
\title{utility to "clean" odd characters in text input that seem to increase risk of LLM failures}
\usage{
cleantxt(x)
}
\arguments{
\item{x}{character(1) text from which non-alphabetic characters like brackets and parentheses
are to be removed}
}
\description{
utility to "clean" odd characters in text input that seem to increase risk of LLM failures
}
\note{
This is speculative; success rates appear to increase with no evident degradation of
content interpretation.
}
\examples{
cleantxt("RNA-seq (paired-end): analysis of #reads [v2]")
}
