% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edam_data}
\alias{edam_data}
\title{snapshot of edam data circa early 2024}
\format{
list with data.frame
}
\usage{
edam_data
}
\description{
snapshot of edam data circa early 2024
}
\keyword{datasets}
