% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edam_formats}
\alias{edam_formats}
\title{snapshot of edam formats circa early 2024}
\format{
list with data.frame
}
\usage{
edam_formats
}
\description{
snapshot of edam formats circa early 2024
}
\keyword{datasets}
