% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edam_operations}
\alias{edam_operations}
\title{snapshot of edam operations circa early 2024}
\format{
list with data.frame
}
\usage{
edam_operations
}
\description{
snapshot of edam operations circa early 2024
}
\keyword{datasets}
