% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edam_topics}
\alias{edam_topics}
\title{snapshot of edam topics circa early 2024}
\format{
list with data.frame
}
\usage{
edam_topics
}
\description{
snapshot of edam topics circa early 2024
}
\keyword{datasets}
