% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edam_graph.R
\name{extract_uri}
\alias{extract_uri}
\title{Helper recursive function to extract 'uri' nodes from JSON document based on edamize}
\usage{
extract_uri(node)
}
\arguments{
\item{node}{a JSON document}
}
\description{
Helper recursive function to extract 'uri' nodes from JSON document based on edamize
}
\note{
From perplexity.ai
}
