% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edamize.R
\name{mkdf}
\alias{mkdf}
\title{simple utility to process output of edamize into a data.frame}
\usage{
mkdf(x)
}
\arguments{
\item{x}{a data.frame as produced by edamize (returned as-is), or a legacy list}
}
\description{
simple utility to process output of edamize into a data.frame
}
\note{
dplyr::distinct is run on the result
}
\examples{
df <- data.frame(uri = "http://edamontology.org/topic_3308",
                 tm  = "Transcriptomics", stringsAsFactors = FALSE)
mkdf(df)  # data.frame input is passed through unchanged
}
