% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksByView.R
\name{pksByViews}
\alias{pksByViews}
\title{helper for package listing}
\usage{
pksByViews(views, type = "BioCsoft", vlist = NULL)
}
\arguments{
\item{views}{character() biocViews node values}

\item{type}{character(1) input to BiocPkgTools::biocPkgList}

\item{vlist}{list() if NULL, use biocPkgList to obtain current biocPkgList, otherwise
a named list with outputs of previous calls to biocPkgList}
}
\description{
helper for package listing
}
\examples{
data("vlist", package="biocEDAM")
pksByViews(views = c("ChIPchip", "ShinyApps"), type="BioCsoft", vlist=vlist)
}
