% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saved_views_2023.18.11}
\alias{saved_views_2023.18.11}
\title{dated set of biocViews}
\format{
An object of class \code{biocPkgList} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2266 rows and 47 columns.
}
\usage{
saved_views_2023.18.11
}
\description{
dated set of biocViews
}
\keyword{datasets}
