% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_bioc.R
\name{tag_bioc}
\alias{tag_bioc}
\title{full workflow to tag a bioconductor package and selected representative content with
EDAM terms}
\usage{
tag_bioc(
  pkgname = "tximeta",
  url =
    "https://bioconductor.org/packages/release/bioc/vignettes/tximeta/inst/doc/tximeta.html"
)
}
\arguments{
\item{pkgname}{character(1)}

\item{url}{URL for representative content, in html or pdf}
}
\value{
a data.frame with biocViews and EDAM suggestions
}
\description{
full workflow to tag a bioconductor package and selected representative content with
EDAM terms
}
\examples{
if (interactive() & nchar(Sys.getenv("OPENAI_API_KEY")>0)) {
ti = tag_bioc()
bs = tag_bioc(pkg="Biostrings", 
  url="https://bioconductor.org/packages/release/bioc/vignettes/Biostrings/inst/doc/Biostrings2Classes.pdf")
library(DT)
ndf = rbind(ti, bs)
datatable(ndf)
}
}
