% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols4_schema.R
\name{TermMappingTable}
\alias{TermMappingTable}
\title{Ellmer type schema for a table of ontology term mappings}
\format{
An ellmer \code{TypeArray} whose items conform to
\code{\link{TermMapping}}.
}
\usage{
TermMappingTable
}
\description{
An array of \code{\link{TermMapping}} objects, suitable for passing as the
\code{type} argument to \code{chat$chat_structured()} when multiple concepts
are to be mapped in a single call.
}
\seealso{
\code{\link{TermMapping}}, \code{\link{ols4_mcp_tools}}
}
