% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_bioc.R
\name{toline}
\alias{toline}
\title{process the output of edamize followed by mkdf to create a data frame with components topic, operation, data, format,
reflecting main elements of EDAM}
\usage{
toline(x)
}
\arguments{
\item{x}{output of mkdf}
}
\value{
a data.frame
}
\description{
process the output of edamize followed by mkdf to create a data frame with components topic, operation, data, format,
reflecting main elements of EDAM
}
\examples{
df <- data.frame(
    uri = c("http://edamontology.org/topic_3308",
            "http://edamontology.org/operation_2238",
            "http://edamontology.org/data_3112",
            "http://edamontology.org/format_3475"),
    tm  = c("Transcriptomics", "Statistical calculation",
            "Gene expression matrix", "TSV"),
    stringsAsFactors = FALSE)
toline(df)
}
