% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vlist}
\alias{vlist}
\title{snapshot of all biocViews}
\format{
An object of class \code{list} of length 4.
}
\usage{
vlist
}
\description{
snapshot of all biocViews
}
\examples{
data(vlist)
sapply(vlist,nrow)
dim(vlist[[1]])
}
\keyword{datasets}
