% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvbrowse.R
\name{bvbrowse}
\alias{bvbrowse}
\title{interactive exploration of biocViews in relation to packages}
\usage{
bvbrowse()
}
\description{
interactive exploration of biocViews in relation to packages
}
\note{
Was made consistent with inst/app code on 3 Nov 2024.
}
