% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_bioc.R
\name{curate_bioc}
\alias{curate_bioc}
\title{use Anh Vu's OpenAI prompting to develop structured metadata about
Bioconductor packages, targeting EDAM ontology and bio.tools schema}
\usage{
curate_bioc(
  packageName = "chromVar",
  devurl =
    "https://raw.githubusercontent.com/GreenleafLab/chromVAR/refs/heads/master/README.md"
)
}
\arguments{
\item{packageName}{character(1) a Bioconductor software package name, its release landing page will be scraped}

\item{devurl}{character(1) a URL for doc originating from the developer}
}
\value{
two python dicts, base_final and edam_processed
}
\description{
use Anh Vu's OpenAI prompting to develop structured metadata about
Bioconductor packages, targeting EDAM ontology and bio.tools schema
}
\examples{
lk = curate_bioc()
str(lk)
}
