% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edam_graph.R
\name{edam_graph}
\alias{edam_graph}
\title{build a graph of EDAM terms deemed relevant to a text}
\usage{
edam_graph(txt, edam_graph, ...)
}
\arguments{
\item{txt}{string, typically describing a software artifact}

\item{edam_graph}{an ontologyIndex ontology-index instance representing EDAM}

\item{\dots}{passed to ontoProc::onto_plot2}
}
\description{
build a graph of EDAM terms deemed relevant to a text
}
\note{
The text is expected to be generated as the 'focused' result
of vig2data; it will then be processed by 'edamize'
}
\examples{
if (interactive()) {
requireNamespace("ontoProc2")
stopifnot(nchar(Sys.getenv("OPENAI_API_KEY"))>0)
eg = readRDS(system.file("rds", "edam_1.25_ontoindex.rds", package="biocEDAM"))
statescoper = readRDS(system.file("rds", "tgac-vumc_StatescopeR.rds", package="biocEDAM"))
edam_graph(statescoper$focused, eg, cex=.3) 
}
}
