% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edamize.R
\name{edamize}
\alias{edamize}
\title{use Anh Vu's OpenAI prompting to develop structured metadata about
Bioconductor packages, targeting EDAM ontology and bio.tools schema}
\usage{
edamize(
  devurl =
    "https://raw.githubusercontent.com/GreenleafLab/chromVAR/refs/heads/master/README.md",
  temp = 0
)
}
\arguments{
\item{devurl}{character(1) a URL for doc originating from the developer}

\item{temp}{numeric(1) temperature setting for openAI chat, see `https://gptcache.readthedocs.io/en/latest/bootcamp/temperature/chat.html`, defaults to 0.0}
}
\value{
two python dicts, base_final and edam_processed
}
\description{
use Anh Vu's OpenAI prompting to develop structured metadata about
Bioconductor packages, targeting EDAM ontology and bio.tools schema
}
\examples{
if (interactive()) {
  key = Sys.getenv("OPENAI_API_KEY")
  if (nchar(key)==0) stop("need to have OPENAI_API_KEY set")
  lk = edamize()
  str(lk)
}
}
