% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_tagger.R
\name{flat_tagger}
\alias{flat_tagger}
\title{simple tagger based on text excerpt and available context from data-frames with edam vocabularies}
\usage{
flat_tagger(txt, nterms = 20, model = "gpt-5", ...)
}
\arguments{
\item{txt}{a text string for analysis}

\item{model}{a string naming an openai model}

\item{\dots}{parameters passed to chat_openai}
}
\description{
simple tagger based on text excerpt and available context from data-frames with edam vocabularies
}
\examples{
if (interactive()) {
txt = "The Voyager package is an R/Bioconductor software designed for exploratory spatial 
data analysis (ESDA) of spatial single-cell omics datasets. It operates on the 
SpatialFeatureExperiment (SFE) S4 class, allowing users to perform a wide range of spatial 
statistical analyses directly within a biological context. Univariate global spatial statistics 
supported include Moran's I for measuring spatial autocorrelation, permutation testing 
for assessing significance, and correlograms for examining spatial correlation structure. 
Bivariate spatial statistics implemented in Voyager comprise Lees L statistic 
and cross variograms for evaluating spatial associations between two 
variables. In addition, Voyager provides tools for multivariate analysis using methods 
such as MULTISPATI PCA, which integrates spatial structure into principal component 
analysis, and Anselins recent multivariate local Gearys C" 
flat_tagger(txt, nterms=12, model="gpt-4o")
}
}
