% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edamize.R
\name{mkdf}
\alias{mkdf}
\title{simple utility to process output of edamize into data.frame}
\usage{
mkdf(x)
}
\arguments{
\item{x}{a list as produced by edamize}
}
\description{
simple utility to process output of edamize into data.frame
}
