% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_bioc.R
\name{toline}
\alias{toline}
\title{process the output of edamize followed by mkdf to create a data frame with components topic, operation, data, format,
reflecting main elements of EDAM}
\usage{
toline(x)
}
\arguments{
\item{x}{output of mkdf}
}
\value{
a data.frame
}
\description{
process the output of edamize followed by mkdf to create a data frame with components topic, operation, data, format,
reflecting main elements of EDAM
}
