% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{given a data.frame with uri and lbl produced by mkdf after edamize,
check against the edam tables in biocEDAM, filtering records to
those with id matching the uri.  the result should give TRUE
for all.equal: all.equal(datf, verify(datf))}
\usage{
verify(datf)
}
\description{
given a data.frame with uri and lbl produced by mkdf after edamize,
check against the edam tables in biocEDAM, filtering records to
those with id matching the uri.  the result should give TRUE
for all.equal: all.equal(datf, verify(datf))
}
\examples{
content = readRDS(system.file("rds/tximetaFocused.rds", package="biocEDAM"))
str(content)
lk = edamize(content$focus)
da = mkdf(lk)
chk = verify(da)
all.equal(da, chk)
}
